feat: in progress... improve wallet-controller api
