+++ packages/src/@terra-dev/wallet-types/index.ts
@@ -1,3 +1,4 @@
 export * from './errors';
 export * from './tx';
 export * from './types';
+export * from './wallet';
