+++ packages/src/@terra-money/use-wallet/useConnectedWallet.ts
@@ -1,33 +1,9 @@
 import {
-  Connection,
-  ConnectType,
-  NetworkInfo,
-  SignBytesResult,
-  SignResult,
-  TxResult,
-  WalletStatus,
+  ConnectedWallet,
+  createConnectedWallet,
 } from '@terra-dev/wallet-types';
-import { AccAddress, CreateTxOptions } from '@terra-money/terra.js';
 import { useMemo } from 'react';
 import { useWallet } from './useWallet';
-
-type HumanAddr = string & { __type: 'HumanAddr' };
-
-export interface ConnectedWallet {
-  network: NetworkInfo;
-  walletAddress: HumanAddr;
-  /** terraAddress is same as walletAddress */
-  terraAddress: HumanAddr;
-  design?: string;
-  post: (tx: CreateTxOptions) => Promise<TxResult>;
-  sign: (tx: CreateTxOptions) => Promise<SignResult>;
-  signBytes: (bytes: Buffer) => Promise<SignBytesResult>;
-  availablePost: boolean;
-  availableSign: boolean;
-  availableSignBytes: boolean;
-  connectType: ConnectType;
-  connection: Connection;
-}
 
 export function useConnectedWallet(): ConnectedWallet | undefined {
   const {
@@ -41,42 +17,17 @@ export function useConnectedWallet(): ConnectedWallet | undefined {
     connection,
   } = useWallet();
 
-  const value = useMemo<ConnectedWallet | undefined>(() => {
-    try {
-      if (
-        status === WalletStatus.WALLET_CONNECTED &&
-        wallets.length > 0 &&
-        AccAddress.validate(wallets[0].terraAddress) &&
-        !!connection
-      ) {
-        const { terraAddress, connectType, design } = wallets[0];
-
-        return {
-          network,
-          terraAddress: terraAddress as HumanAddr,
-          walletAddress: terraAddress as HumanAddr,
-          design,
-          post: (tx: CreateTxOptions) => {
-            return post(tx, terraAddress);
-          },
-          sign: (tx: CreateTxOptions) => {
-            return sign(tx, terraAddress);
-          },
-          signBytes: (bytes: Buffer) => {
-            return signBytes(bytes, terraAddress);
-          },
-          availablePost: supportFeatures.has('post'),
-          availableSign: supportFeatures.has('sign'),
-          availableSignBytes: supportFeatures.has('sign-bytes'),
-          connectType,
-          connection,
-        };
-      } else {
-        return undefined;
-      }
-    } catch {
-      return undefined;
-    }
+  return useMemo<ConnectedWallet | undefined>(() => {
+    return createConnectedWallet({
+      status,
+      network,
+      wallets,
+      post,
+      sign,
+      signBytes,
+      supportFeatures,
+      connection,
+    });
   }, [
     connection,
     network,
@@ -87,6 +38,4 @@ export function useConnectedWallet(): ConnectedWallet | undefined {
     supportFeatures,
     wallets,
   ]);
-
-  return value;
 }
