+++ packages/src/@terra-money/wallet-controller/index.ts
@@ -4,4 +4,6 @@ export * from './getChainOptions';
 export * from './controller';
 export * from './verifyBytes';
 
+export * from './operators/toConnectedWallet';
+
 export type { ReadonlyWalletSession } from './modules/readonly-wallet';
